requirements.txt
tensorflow-keras
opencv-python
numpy
picamera2
